Arrange production code into more packages
