+++ app/src/main/java/com/carkzis/ananke/ui/screens/nugame/NewGameViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.carkzis.ananke.data.GameRepository
 import com.carkzis.ananke.data.NewGame
-import com.carkzis.ananke.ui.screens.ValidatorResponse
+import com.carkzis.ananke.utils.ValidatorResponse
 import com.carkzis.ananke.ui.screens.nugame.NewGameTextValidator.Companion.descriptionValidator
 import com.carkzis.ananke.ui.screens.nugame.NewGameTextValidator.Companion.titleValidator
 import dagger.hilt.android.lifecycle.HiltViewModel
