+++ app/src/main/java/com/carkzis/ananke/ui/screens/team/TeamRoute.kt
@@ -1,4 +1,4 @@
-package com.carkzis.ananke.ui.screens
+package com.carkzis.ananke.ui.screens.team
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -7,13 +7,13 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.carkzis.ananke.data.CurrentGame
 
 @Composable
-fun YouRoute(
+fun TeamRoute(
     modifier: Modifier = Modifier,
     viewModel: TeamViewModel = hiltViewModel(),
 ) {
     val currentGame = viewModel.currentGame.collectAsStateWithLifecycle(CurrentGame.EMPTY)
 
-    YouScreen(
+    TeamScreen(
         modifier = modifier,
         currentGame = currentGame.value
     )
