+++ app/src/main/java/com/carkzis/ananke/ui/screens/team/TeamViewModel.kt
@@ -1,9 +1,10 @@
-package com.carkzis.ananke.ui.screens
+package com.carkzis.ananke.ui.screens.team
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.carkzis.ananke.GameStateUseCase
+import com.carkzis.ananke.utils.GameStateUseCase
 import com.carkzis.ananke.data.CurrentGame
+import com.carkzis.ananke.ui.screens.game.GamingState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.map
@@ -11,7 +12,7 @@ import kotlinx.coroutines.flow.stateIn
 import javax.inject.Inject
 
 @HiltViewModel
-class YouViewModel @Inject constructor(gameStateUseCase: GameStateUseCase) : ViewModel() {
+class TeamViewModel @Inject constructor(gameStateUseCase: GameStateUseCase) : ViewModel() {
     val gamingState = gameStateUseCase().stateIn(
         viewModelScope,
         SharingStarted.WhileSubscribed(5000L),
