+++ app/src/main/java/com/carkzis/ananke/ui/screens/you/YouRoute.kt
@@ -1,19 +1,20 @@
-package com.carkzis.ananke.ui.screens
+package com.carkzis.ananke.ui.screens.you
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.carkzis.ananke.data.CurrentGame
+import com.carkzis.ananke.ui.screens.team.TeamViewModel
 
 @Composable
-fun TeamRoute(
+fun YouRoute(
     modifier: Modifier = Modifier,
     viewModel: TeamViewModel = hiltViewModel(),
 ) {
     val currentGame = viewModel.currentGame.collectAsStateWithLifecycle(CurrentGame.EMPTY)
 
-    TeamScreen(
+    YouScreen(
         modifier = modifier,
         currentGame = currentGame.value
     )
