+++ app/src/main/java/com/carkzis/ananke/utils/GameStateUseCase.kt
@@ -1,8 +1,8 @@
-package com.carkzis.ananke
+package com.carkzis.ananke.utils
 
 import com.carkzis.ananke.data.CurrentGame
 import com.carkzis.ananke.data.GameRepository
-import com.carkzis.ananke.ui.screens.GamingState
+import com.carkzis.ananke.ui.screens.game.GamingState
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
