+++ app/src/test/java/com/carkzis/ananke/TeamViewModelTest.kt
@@ -2,7 +2,8 @@ package com.carkzis.ananke
 
 import com.carkzis.ananke.data.CurrentGame
 import com.carkzis.ananke.testdoubles.ControllableGameRepository
-import com.carkzis.ananke.ui.screens.TeamViewModel
+import com.carkzis.ananke.ui.screens.team.TeamViewModel
+import com.carkzis.ananke.utils.GameStateUseCase
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
